Revert "Moved to CLion/cmake"
This reverts commit 32f5372bcdb1f7f34bc4d2aced361469975cd27a.

+++ src/PEProcess.cpp
@@ -184,7 +184,7 @@ std::string PEProcess::ReadMemoryString(uintptr_t offset, int length, int* resul
 {
     LPVOID address = CheckAddress(offset, directAddress);
 
-    SIZE_T totalRead = 0;
+    int totalRead = 0;
     std::string str;
     do
     {
